Split multi-shard replica deletion into separate v2 API

+++ solr/core/src/java/org/apache/solr/handler/admin/api/DeleteReplicaAPI.java
@@ -31,10 +31,12 @@ import static org.apache.solr.common.params.CoreAdminParams.DELETE_INSTANCE_DIR;
 import static org.apache.solr.common.params.CoreAdminParams.REPLICA;
 import static org.apache.solr.security.PermissionNameProvider.Name.COLL_EDIT_PERM;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.HashMap;
 import java.util.Map;
 import javax.inject.Inject;
 import javax.ws.rs.DELETE;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.QueryParam;
@@ -44,17 +46,17 @@ import org.apache.solr.common.params.CollectionParams;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.core.CoreContainer;
 import org.apache.solr.handler.api.V2ApiUtils;
+import org.apache.solr.jersey.JacksonReflectMapWriter;
 import org.apache.solr.jersey.PermissionName;
 import org.apache.solr.jersey.SubResponseAccumulatingJerseyResponse;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
 
 /**
- * V2 APIs for deleting one or more existing replicas from a shard.
+ * V2 APIs for deleting one or more existing replicas from one or more shards.
  *
  * <p>These APIs are analogous to the v1 /admin/collections?action=DELETEREPLICA command.
  */
-@Path("/collections/{collectionName}/shards/{shardName}/replicas")
 public class DeleteReplicaAPI extends AdminAPIBase {
 
   @Inject
@@ -66,7 +68,7 @@ public class DeleteReplicaAPI extends AdminAPIBase {
   }
 
   @DELETE
-  @Path("/{replicaName}")
+  @Path("/collections/{collectionName}/shards/{shardName}/replicas/{replicaName}")
   @PermissionName(COLL_EDIT_PERM)
   public SubResponseAccumulatingJerseyResponse deleteReplicaByName(
       @PathParam("collectionName") String collectionName,
@@ -105,6 +107,7 @@ public class DeleteReplicaAPI extends AdminAPIBase {
   }
 
   @DELETE
+  @Path("/collections/{collectionName}/shards/{shardName}/replicas")
   @PermissionName(COLL_EDIT_PERM)
   public SubResponseAccumulatingJerseyResponse deleteReplicasByCount(
       @PathParam("collectionName") String collectionName,
@@ -142,6 +145,64 @@ public class DeleteReplicaAPI extends AdminAPIBase {
     return response;
   }
 
+  public static class ScaleCollectionRequestBody implements JacksonReflectMapWriter {
+    public @JsonProperty(value = COUNT_PROP, required = true) Integer numToDelete;
+    public @JsonProperty(FOLLOW_ALIASES) Boolean followAliases;
+    public @JsonProperty(DELETE_INSTANCE_DIR) Boolean deleteInstanceDir;
+    public @JsonProperty(DELETE_DATA_DIR) Boolean deleteDataDir;
+    public @JsonProperty(DELETE_INDEX) Boolean deleteIndex;
+    public @JsonProperty(ONLY_IF_DOWN) Boolean onlyIfDown;
+    public @JsonProperty(ASYNC) String asyncId;
+
+    public static ScaleCollectionRequestBody fromV1Params(SolrParams v1Params) {
+      final var requestBody = new ScaleCollectionRequestBody();
+      requestBody.numToDelete = v1Params.getInt(COUNT_PROP);
+      requestBody.followAliases = v1Params.getBool(FOLLOW_ALIASES);
+      requestBody.deleteInstanceDir = v1Params.getBool(DELETE_INSTANCE_DIR);
+      requestBody.deleteDataDir = v1Params.getBool(DELETE_DATA_DIR);
+      requestBody.deleteIndex = v1Params.getBool(DELETE_INDEX);
+      requestBody.onlyIfDown = v1Params.getBool(ONLY_IF_DOWN);
+      requestBody.asyncId = v1Params.get(ASYNC);
+
+      return requestBody;
+    }
+  }
+
+  @PUT
+  @Path("/collections/{collectionName}/scale")
+  @PermissionName(COLL_EDIT_PERM)
+  public SubResponseAccumulatingJerseyResponse deleteReplicasByCountAllShards(
+      @PathParam("collectionName") String collectionName, ScaleCollectionRequestBody requestBody)
+      throws Exception {
+    final var response = instantiateJerseyResponse(SubResponseAccumulatingJerseyResponse.class);
+    if (requestBody == null) {
+      throw new SolrException(BAD_REQUEST, "Request body is required but missing");
+    }
+    ensureRequiredParameterProvided(COLLECTION_PROP, collectionName);
+    ensureRequiredParameterProvided(COUNT_PROP, requestBody.numToDelete);
+    fetchAndValidateZooKeeperAwareCoreContainer();
+    recordCollectionForLogAndTracing(collectionName, solrQueryRequest);
+
+    final ZkNodeProps remoteMessage =
+        createRemoteMessage(
+            collectionName,
+            null,
+            null,
+            requestBody.numToDelete,
+            requestBody.followAliases,
+            requestBody.deleteInstanceDir,
+            requestBody.deleteDataDir,
+            requestBody.deleteIndex,
+            requestBody.onlyIfDown,
+            requestBody.asyncId);
+    submitRemoteMessageAndHandleResponse(
+        response,
+        CollectionParams.CollectionAction.DELETEREPLICA,
+        remoteMessage,
+        requestBody.asyncId);
+    return response;
+  }
+
   public static ZkNodeProps createRemoteMessage(
       String collectionName,
       String shardName,
@@ -156,7 +217,7 @@ public class DeleteReplicaAPI extends AdminAPIBase {
     final Map<String, Object> remoteMessage = new HashMap<>();
     remoteMessage.put(QUEUE_OPERATION, CollectionParams.CollectionAction.DELETEREPLICA.toLower());
     remoteMessage.put(COLLECTION_PROP, collectionName);
-    remoteMessage.put(SHARD_ID_PROP, shardName);
+    insertIfNotNull(remoteMessage, SHARD_ID_PROP, shardName);
     insertIfNotNull(remoteMessage, REPLICA, replicaName);
     insertIfNotNull(remoteMessage, COUNT_PROP, numReplicasToDelete);
     insertIfNotNull(remoteMessage, FOLLOW_ALIASES, followAliases);
@@ -175,7 +236,7 @@ public class DeleteReplicaAPI extends AdminAPIBase {
       SolrQueryResponse solrQueryResponse)
       throws Exception {
     final var v1Params = solrQueryRequest.getParams();
-    v1Params.required().check(COLLECTION_PROP, SHARD_ID_PROP);
+    v1Params.required().check(COLLECTION_PROP);
 
     final var deleteReplicaApi =
         new DeleteReplicaAPI(coreContainer, solrQueryRequest, solrQueryResponse);
@@ -185,7 +246,7 @@ public class DeleteReplicaAPI extends AdminAPIBase {
 
   private static SubResponseAccumulatingJerseyResponse invokeApiMethod(
       DeleteReplicaAPI deleteReplicaApi, SolrParams v1Params) throws Exception {
-    if (v1Params.get(REPLICA) != null) {
+    if (v1Params.get(REPLICA) != null && v1Params.get(SHARD_ID_PROP) != null) {
       return deleteReplicaApi.deleteReplicaByName(
           v1Params.get(COLLECTION_PROP),
           v1Params.get(SHARD_ID_PROP),
@@ -196,7 +257,8 @@ public class DeleteReplicaAPI extends AdminAPIBase {
           v1Params.getBool(DELETE_INDEX),
           v1Params.getBool(ONLY_IF_DOWN),
           v1Params.get(ASYNC));
-    } else if (v1Params.get(COUNT_PROP) != null) {
+    } else if (v1Params.get(SHARD_ID_PROP) != null
+        && v1Params.get(COUNT_PROP) != null) { // Delete 'N' replicas
       return deleteReplicaApi.deleteReplicasByCount(
           v1Params.get(COLLECTION_PROP),
           v1Params.get(SHARD_ID_PROP),
@@ -207,6 +269,9 @@ public class DeleteReplicaAPI extends AdminAPIBase {
           v1Params.getBool(DELETE_INDEX),
           v1Params.getBool(ONLY_IF_DOWN),
           v1Params.get(ASYNC));
+    } else if (v1Params.get(COUNT_PROP) != null) {
+      return deleteReplicaApi.deleteReplicasByCountAllShards(
+          v1Params.get(COLLECTION_PROP), ScaleCollectionRequestBody.fromV1Params(v1Params));
     } else {
       throw new SolrException(
           BAD_REQUEST,
